coupon-backend/requirements.txt
asgiref==3.3.1
click==7.1.2
Django==3.1.3
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
h11==0.11.0
Pillow==8.0.1
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
uvicorn==0.12.3
uWSGI==2.0.19.1
